z_formats::settings_format
Struct Parsed

pub struct Parsed<'s> {
    _0: Parsed<'s, [Segment<'s>; 2]>,
}

Fields
_0: Parsed<'s, [Segment<'s>; 2]>
Implementations

impl Parsed<'_>

pub fn user_id(&self) -> &keyexpr

Get the parsed value for user_id.

This value is guaranteed to be a valid key expression intersecting with *

pub fn setting(&self) -> Option<&keyexpr>

Get the parsed value for setting.

This value is guaranteed to be a valid key expression intersecting with ** Since the pattern is **, this may return None if the pattern didn’t consume any chunks.