z_formats
Module file_format

The module associated with the user_id/${user_id:*}/file/${file:*/**} format, it contains:

Format, a zero-sized type that represents your format.

formatter(), a function that constructs a Formatter specialized for your format:

for every spec in your format, Formatter will have a method named after the spec’s id that lets you set a value for that field of your format. These methods will return Result<&mut Formatter, FormatError>.

parse(target: &keyexpr) -> ZResult<Parsed<'_>> will parse the provided key expression according to your format. Just like KeFormat::parse, parsing is lazy: each field will match the smallest subsection of your target that is included in its pattern.

like Formatter, Parsed will have a method named after each spec’s id that returns &keyexpr; except for specs whose pattern was **, these will return an Option<&keyexpr>, where None signifies that the pattern was matched by an empty list of chunks.

Structs

Format

The user_id/${user_id:*}/file/${file:*/**} format, as a zero-sized-type.

Formatter

And instance of a formatter for user_id/${user_id:*}/file/${file:*/**}.

Parsed

Constants

FORMAT_INNER 🔒

Functions

formatter

parse